+++ src/Palya/Random/palya_feltoltes/Random.h
@@ -6,7 +6,11 @@
 #define MINESWEEPER_RANDOM_H
 #include <random>
 
-int random(int a, int b);
+/// @brief A két paraméterül kapott szám között generál egy random számot, azt adja vissza
+/// \param a - az alsóhatár
+/// \param b - a felsőhatár
+/// \return A generált szám
+size_t random(size_t a, size_t b);
 
 
 #endif //MINESWEEPER_RANDOM_H
